added test to check serialization of opt state, moved test to test_models
